kvserver,allocator: move replica planner into pkg
In order to isolate the decision logic further and clean up the
replicate queue file, this commit moves the `ReplicationPlanner` and
associated logic into the `plan` pkg.

This commit makes no changes to behavior.

Part of #90141

Release note: None
