+++ pkg/BUILD.bazel
@@ -1243,6 +1243,7 @@ GO_TARGETS = [
     "//pkg/kv/kvserver/allocator/allocatorimpl:allocatorimpl",
     "//pkg/kv/kvserver/allocator/allocatorimpl:allocatorimpl_test",
     "//pkg/kv/kvserver/allocator/load:load",
+    "//pkg/kv/kvserver/allocator/plan:plan",
     "//pkg/kv/kvserver/allocator/storepool:storepool",
     "//pkg/kv/kvserver/allocator/storepool:storepool_test",
     "//pkg/kv/kvserver/allocator:allocator",
@@ -2737,6 +2738,7 @@ GET_X_DATA_TARGETS = [
     "//pkg/kv/kvserver/allocator/allocator2:get_x_data",
     "//pkg/kv/kvserver/allocator/allocatorimpl:get_x_data",
     "//pkg/kv/kvserver/allocator/load:get_x_data",
+    "//pkg/kv/kvserver/allocator/plan:get_x_data",
     "//pkg/kv/kvserver/allocator/storepool:get_x_data",
     "//pkg/kv/kvserver/apply:get_x_data",
     "//pkg/kv/kvserver/asim:get_x_data",
