+++ pkg/kv/kvserver/liveness/BUILD.bazel
@@ -46,6 +46,7 @@ go_test(
     deps = [
         "//pkg/base",
         "//pkg/kv/kvserver",
+        "//pkg/kv/kvserver/allocator/plan",
         "//pkg/kv/kvserver/allocator/storepool",
         "//pkg/kv/kvserver/liveness/livenesspb",
         "//pkg/roachpb",
