+++ pkg/kv/kvserver/liveness/client_test.go
@@ -20,6 +20,7 @@ import (
 
 	"github.com/cockroachdb/cockroach/pkg/base"
 	"github.com/cockroachdb/cockroach/pkg/kv/kvserver"
+	"github.com/cockroachdb/cockroach/pkg/kv/kvserver/allocator/plan"
 	"github.com/cockroachdb/cockroach/pkg/kv/kvserver/allocator/storepool"
 	"github.com/cockroachdb/cockroach/pkg/kv/kvserver/liveness"
 	"github.com/cockroachdb/cockroach/pkg/kv/kvserver/liveness/livenesspb"
@@ -144,7 +145,7 @@ func TestNodeLivenessStatusMap(t *testing.T) {
 			Store: &kvserver.StoreTestingKnobs{
 				// Disable replica rebalancing to ensure that the liveness range
 				// does not get out of the first node (we'll be shutting down nodes).
-				ReplicaPlannerKnobs: kvserver.ReplicaPlannerTestingKnobs{
+				ReplicaPlannerKnobs: plan.ReplicaPlannerTestingKnobs{
 					DisableReplicaRebalancing: true,
 				},
 				// Disable LBS because when the scan is happening at the rate it's happening
