+++ pkg/server/status_test.go
@@ -36,6 +36,7 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/jobs/jobspb"
 	"github.com/cockroachdb/cockroach/pkg/keys"
 	"github.com/cockroachdb/cockroach/pkg/kv/kvserver"
+	"github.com/cockroachdb/cockroach/pkg/kv/kvserver/allocator/plan"
 	"github.com/cockroachdb/cockroach/pkg/roachpb"
 	"github.com/cockroachdb/cockroach/pkg/rpc"
 	"github.com/cockroachdb/cockroach/pkg/security/username"
@@ -340,7 +341,7 @@ func startServer(t *testing.T) *TestServer {
 			Store: &kvserver.StoreTestingKnobs{
 				// Now that we allow same node rebalances, disable it in these tests,
 				// as they dont expect replicas to move.
-				ReplicaPlannerKnobs: kvserver.ReplicaPlannerTestingKnobs{
+				ReplicaPlannerKnobs: plan.ReplicaPlannerTestingKnobs{
 					DisableReplicaRebalancing: true,
 				},
 			},
