+++ pkg/server/structlogging/BUILD.bazel
@@ -33,6 +33,7 @@ go_test(
         "//pkg/base",
         "//pkg/ccl",
         "//pkg/kv/kvserver",
+        "//pkg/kv/kvserver/allocator/plan",
         "//pkg/roachpb",
         "//pkg/security/securityassets",
         "//pkg/security/securitytest",
